+++ app/utils/download_products.py
@@ -11,57 +11,19 @@ import numpy as np
 import subprocess
 from app.models.execution import INPUT_PRODUCTS_PATH
 
-# 1. Create an account with CDDIS
-# 2. Create a file named "cddis.env" in the same directory as this script
-# 3. Add EMAIL=<your_account_email> to the "cddis.env" file
-
-# Ensure account is registered and in this file :)
-load_dotenv(Path(__file__).parent / "cddis.env")
-
-
-def retrieve_all_cddis_types(reference_start: GPSDate) -> list[str]:
-    """
-    Retrieve all CDDIS data types for a given GPS Week.
-
-    :param reference_start: The datetime of the GPS Week to retrieve.
-    :param timespan: The duration for which to retrieve data.
-    :return:
-    """
-    ftp_tls = FTP_TLS(host="gdc.cddis.eosdis.nasa.gov", user="anonymous", passwd=os.getenv("EMAIL"), timeout=60)
-    ftp_tls.prot_p()  # Secures the TLS connection, mandatory for CDDIS
-    files = None
-    try:
-        ftp_tls.cwd(f"gnss/products/{reference_start.gpswk}")
-        files = ftp_tls.nlst()
-    except ftplib.all_errors as e:
-        print("Error getting file list", e)
-    return files
-
-def create_cddis_file(filepath: Path, reference_start: GPSDate) -> None:
-    """
-    Create a file named "CDDIS.list" with CDDIS data types for a given reference start time.
-
-    :param filepath: The path to the directory where the file will be created.
-    :param reference_start: The start time for the data retrieval.
-    """
-    data = retrieve_all_cddis_types(reference_start)
-    with open(filepath.joinpath("../models/CDDIS.list"), "w") as f:
-        for d in data:
-            try:
-                time = datetime.strptime(d.split("_")[1], "%Y%j%H%M")
-                f.write(f"{d} {time} \n")
-            except IndexError:
-                data.remove(d)
-
 def download_ppp_products(inputs) -> bool:
     """Download PPP products using the CDDIS_handler and auto_download_PPP script"""
-    start_datetime  = inputs.start_epoch
-    end_datetime    = inputs.end_epoch
+    start_datetime  = inputs.start_epoch.replace(" ", "_")
+    end_datetime    = inputs.end_epoch.replace(" ", "_")
 
     cddis = CDDIS_Handler(end_datetime)
 
     analysis_center = inputs.ppp_provider.upper()
     project_type, solution_type = cddis.get_optimal_project_solution_tuple(analysis_center)
+
+    if not project_type or not solution_type:
+        print(f"No valid products available for {analysis_center}")
+        return False
 
     try:
         download_static_products(start_datetime, end_datetime)
@@ -124,7 +86,3 @@ def download_dynamic_products(
     print(f"Downloading dynamic PPP products for {analysis_center}...")
     result = subprocess.run(command, check=True, capture_output=True, text=True)
     print("Dynamic products downloaded successfully")
-
-if __name__ == "__main__":
-    start_time = GPSDate(np.datetime64(datetime(2023, 10, 1, 0, 0)))
-    create_cddis_file(Path(__file__).parent, start_time)
